Fixed a gameplay issue and updated credits
I made it so that now whenever you activate a checkpoint it also saves the game instead of only when you collect presents as a quality of life feature. If I didn't do this then for example in the final section of the game if you were to quit and load then you would have to navigate through tensegrity again to get back (you wouldn't need to collect the presents again but it's still annoying). I also updated the credits to include me since I programmed the save system, and it also feels nice to think I helped play a part in making this game. Especially since I was ill during the main development of it.